Keep MEMORY_PER_WORKER in Int64 so 32-bit Julia doesn't overflow
3 * 2^30 overflows a 32-bit Int, which made default_max_workers() return 1 on x86.

+++ src/TestItemRuns.jl
@@ -75,7 +75,7 @@ export default_max_workers
 """
 Memory budgeted per test process by [`default_max_workers`](@ref), in bytes (3 GiB).
 """
-const MEMORY_PER_WORKER = 3 * 2^30
+const MEMORY_PER_WORKER = Int64(3) * 2^30  # Int64: 3 GiB overflows a 32-bit Int
 
 """
     default_max_workers(; total_memory=Sys.total_memory(), cpu_threads=Sys.CPU_THREADS) -> Int
